--- a/paket.lock
+++ b/paket.lock
@@ -1,6 +1,7 @@
 NUGET
   remote: https://www.nuget.org/api/v2
   specs:
+    BenchmarkDotNet (0.9.5)
     FParsec (1.0.2)
     FSharp.Core (4.0.0.1)
     Newtonsoft.Json (8.0.3)