--- a/tests/FSharp.Data.GraphQL.Benchmarks/Prolog.fs
+++ b/tests/FSharp.Data.GraphQL.Benchmarks/Prolog.fs
@@ -1,0 +1,15 @@
+﻿/// The MIT License (MIT)
+/// Copyright (c) 2016 Bazinga Technologies Inc
+
+[<AutoOpen>]
+module FSharp.Data.GraphQL.BenchmarkProlog
+
+open BenchmarkDotNet.Configs
+open BenchmarkDotNet.Columns
+open BenchmarkDotNet.Exporters
+
+type GraphQLBenchConfig() as this= 
+    inherit ManualConfig()
+    do
+        this.Add(StatisticColumn.Mean, StatisticColumn.Min, StatisticColumn.Max, StatisticColumn.OperationsPerSecond)
+        this.Add(MarkdownExporter.GitHub)