- Added performance benchmarking tests.
- Additionally removed multiple optional parameters from `Schema` in favor of single optional `SchemaConfig` record.

## Initial benchmarks results

They don't look impressive so we definitely have a room for improvements here:

### Parsing benchmark
```ini

BenchmarkDotNet=v0.9.5.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i5-4430 CPU @ 3.00GHz, ProcessorCount=4
Frequency=2929689 ticks, Resolution=341.3332 ns, Timer=TSC
HostCLR=MS.NET 4.0.30319.42000, Arch=32-bit RELEASE
JitModules=clrjit-v4.6.1055.0

Type=ParsingBenchmark  Mode=Throughput  

```
            Method |      Median |    StdDev |        Mean |         Min |         Max |     Op/s |
------------------ |------------ |---------- |------------ |------------ |------------ |--------- |
  ParseSimpleQuery |  33.8274 us | 0.3540 us |  33.7482 us |  33.0816 us |  34.1611 us | 29631.22 |
 ParseComplexQuery | 577.0307 us | 8.9946 us | 578.8314 us | 566.7890 us | 594.1037 us |  1727.62 |

### Querying benchmark
```ini

BenchmarkDotNet=v0.9.5.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i5-4430 CPU @ 3.00GHz, ProcessorCount=4
Frequency=2929689 ticks, Resolution=341.3332 ns, Timer=TSC
HostCLR=MS.NET 4.0.30319.42000, Arch=32-bit RELEASE
JitModules=clrjit-v4.6.1055.0

Type=SimpleExecutionBenchmark  Mode=Throughput  

```
                       Method |      Median |     StdDev |        Mean |         Min |         Max |     Op/s |
----------------------------- |------------ |----------- |------------ |------------ |------------ |--------- |
 BenchmarkSimpleQueryUnparsed |  71.0288 us |  0.8807 us |  71.3812 us |  70.0700 us |  72.8218 us |  14009.3 |
   BenchmarkSimpleQueryParsed |  24.4913 us |  0.6096 us |  24.5193 us |  23.8519 us |  26.7201 us | 40784.22 |
   BenchmarkFlatQueryUnparsed | 125.2367 us |  3.0207 us | 126.3752 us | 124.3089 us | 135.1953 us |  7912.95 |
     BenchmarkFlatQueryParsed |  52.4176 us |  2.2771 us |  52.3338 us |  48.0172 us |  55.2310 us | 19108.12 |
 BenchmarkNestedQueryUnparsed | 394.4258 us |  9.6195 us | 398.5617 us | 390.9118 us | 427.4011 us |  2509.02 |
   BenchmarkNestedQueryParsed | 170.9940 us | 13.3266 us | 175.4107 us | 165.8474 us | 223.2751 us |  5700.91 |